refactor(Player): Player에 bettingAmount 필드 type 변경 및 기능 추가
- 배팅 정보 DTO 생성

+++ src/main/java/model/participant/state/EarningCalculator.java
@@ -1,5 +1,5 @@
 package model.participant.state;
 
 public interface EarningCalculator {
-    int calculate(int bettingAmount);
+    long calculate(long bettingAmount);
 }
